requirements.txt
pandas
numpy
openai
python-dotenv
matplotlib
yfinance
ccxt
